Made leaderboardFragment class as well as matching fragment xml Custom listview adapter and listview for leaderboard
