+++ SuperQR/.idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/SuperQR/CMPUT301_GroupProject/SuperQR/app/src/main/res/layout/activity_main.xml" value="0.67" />
+        <entry key="..\:/SuperQR/CMPUT301_GroupProject/SuperQR/app/src/main/res/layout/fragment_leaderboard.xml" value="0.5" />
         <entry key="..\:/SuperQR/CMPUT301_GroupProject/SuperQR/app/src/main/res/layout/leaderboard_content.xml" value="0.1" />
         <entry key="..\:/SuperQR/CMPUT301_GroupProject/SuperQR/app/src/main/res/layout/main_activity.xml" value="0.33" />
       </map>
