+++ SuperQR/app/src/main/java/com/example/superqr/leaderboardFragment.java
@@ -1,0 +1,98 @@
+package com.example.superqr;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.ListView;
+
+import java.util.ArrayList;
+import java.util.Collections;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link leaderboardFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class leaderboardFragment extends Fragment {
+
+    // TODO: Rename parameter arguments, choose names that match
+    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+    private static final String ARG_PARAM1 = "param1";
+    private static final String ARG_PARAM2 = "param2";
+
+    // TODO: Rename and change types of parameters
+    private String mParam1;
+    private String mParam2;
+    private ListView leaderboardList;
+    private ArrayAdapter<Player> playerAdapter;
+    private ArrayList<Player> playerList;
+
+    public leaderboardFragment() {
+        // Required empty public constructor
+    }
+
+    /**
+     * Use this factory method to create a new instance of
+     * this fragment using the provided parameters.
+     *
+     * @param param1 Parameter 1.
+     * @param param2 Parameter 2.
+     * @return A new instance of fragment leaderboardFragment.
+     */
+    // TODO: Rename and change types and number of parameters
+    public static leaderboardFragment newInstance(String param1, String param2) {
+        leaderboardFragment fragment = new leaderboardFragment();
+        Bundle args = new Bundle();
+        args.putString(ARG_PARAM1, param1);
+        args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            mParam1 = getArguments().getString(ARG_PARAM1);
+            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
+        View view = LayoutInflater.from(getActivity()).inflate(R.layout.fragment_leaderboard,null);
+        leaderboardList = view.findViewById(R.id.leaderboard_list);
+        playerList = new ArrayList<>();
+
+        playerList.add(new Player("Bri",450));
+        playerList.add(new Player("Chris",650));
+        playerList.add(new Player("Rob",250));
+        playerList.add(new Player("Scarlet",500));
+
+
+        Collections.sort(playerList);
+        playerAdapter = new LeaderboardListView(getActivity(), playerList);
+        leaderboardList.setAdapter(playerAdapter);
+
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View view = LayoutInflater.from(getActivity()).inflate(R.layout.fragment_leaderboard,null);
+        leaderboardList = view.findViewById(R.id.leaderboard_list);
+        playerList = new ArrayList<>();
+
+        playerList.add(new Player("Bri",450));
+        playerList.add(new Player("Chris",650));
+        playerList.add(new Player("Rob",250));
+        playerList.add(new Player("Scarlet",500));
+
+
+        Collections.sort(playerList);
+        playerAdapter = new LeaderboardListView(getActivity(), playerList);
+        leaderboardList.setAdapter(playerAdapter);
+        return view;
+    }
+}
