+++ SuperQR/app/src/main/res/layout/leaderboard_content.xml
@@ -1,4 +1,4 @@
-<RelativeLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/RelativeLayout01"
     android:layout_width="fill_parent"
@@ -8,17 +8,32 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:id="@+id/userId"
+        android:layout_weight="1"
+        android:layout_gravity="center"
+        android:id="@+id/Rank"
         android:padding="20dp"
         android:textSize="20sp"
+        android:textAlignment="center"
         />
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentRight="true"
+        android:layout_weight="2"
+        android:id="@+id/userId"
+        android:padding="20dp"
+        android:layout_gravity="center"
+        android:textSize="20sp"
+        android:textAlignment="center"
+        />
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_weight="1"
+        android:layout_height="wrap_content"
         android:id="@+id/score"
         android:padding="20dp"
+        android:layout_gravity="center"
+        android:textAlignment="center"
         android:textSize="20sp"/>
 
-</RelativeLayout>
+
+</LinearLayout>
